envoy: stream SVIDs updates to cilium-proxy
This commit implements all the logic to stream SVIDs to cilium-proxy.
An SVID is pushed when the endpoint gets a new identity or when the SVID
is rotated.

Certificates from spire are received in DER format, those are converted
to PEM before sending. TODO: check which encoding is more efficient
and if we can use DER directly in the cilium-proxy to avoid the
conversion.