spiffe: automatically recover from errors
This commit improves error handling by:

- Restablishing conneciton to spire agent when lost: cilium-agent
tries to reconnect and after the connection is successful it adds
the pods to the watched set again.

- Tracking the expiration time of the spiffe-based labels. Spiffe-based
labels are removed from the endpoint if they are not renewed after they
expire. This is useful to handle situation where the spire control plane
is down and cilium-agent doesn't know the status of those entries.