hubble: Check for flow type in TCP flag flow filter
Cilium agent crashes when an L7/HTTP flow is passed to TCP flag flow
filter (`filterByTCPFlags`). This is because HTTP flows will have
some L4/TCP info such as src/dst port in the flow struct, but will not
contain TCP flags.

Added checks to look for TCP flags in the flow struct only if the flow
is of type `FlowType_L3_L4`

Signed-off-by: Wazir Ahmed <wazir@accuknox.com>